dspy-trainingv2/requirements.txt
# DSPy OpenCode Optimizer Requirements

# Core dependencies
dspy-ai>=2.4.0
pydantic>=2.0.0
typer[all]>=0.9.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional dependencies for different LM providers
openai>=1.0.0
anthropic>=0.21.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
